Desenvolvendo a entidade Categoria

+++ Backend/sgc/src/main/java/br/com/turma/sgc/repository/CategoriaRepository.java
@@ -4,9 +4,6 @@ import br.com.turma.sgc.domain.Categoria;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.util.List;
-
 @Repository
-public interface CategoriaRepository extends JpaRepository<Categoria, Integer> {
-
+public interface CategoriaRepository extends JpaRepository<Categoria, Long> {
 }
